+++ R/aaa.R
@@ -1,3 +1,0 @@
-#' Internal package environment for usmapdata.
-#' @keywords internal
-.pkg_env <- new.env(parent = emptyenv())